LLM_Resume_Parser_App/requirements.txt
# Python Version 3.9.0

langchain==0.2.10
langchain_core==0.2.22
langchain_groq==0.1.6
pydantic==2.8.2
email_validator==2.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
streamlit==1.36.0
fastapi
uvicorn
aiofiles
python-multipart